# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Attendance Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      parameters:
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: semester
          schema:
            type: string
        - in: query
          name: facultyId
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a new subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ATTENDANCE ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Subject:
      type: object
      required: [id, subjectName, subjectCode, department, semester, createdAt]
      properties:
        id:
          type: integer
        subjectName:
          type: string
        subjectCode:
          type: string
        department:
          type: string
        semester:
          type: string
        facultyId:
          type: ["integer", "null"]
        facultyName:
          type: ["string", "null"]
        createdAt:
          type: string
    SubjectInput:
      type: object
      required: [subjectName, subjectCode, department, semester]
      properties:
        subjectName:
          type: string
        subjectCode:
          type: string
        department:
          type: string
        semester:
          type: string
        facultyId:
          type: ["integer", "null"]
    SubjectUpdate:
      type: object
      properties:
        subjectName:
          type: string
        subjectCode:
          type: string
        department:
          type: string
        semester:
          type: string
        facultyId:
          type: ["integer", "null"]

    # ─── ATTENDANCE ──────────────────────────────────────────────────────────
...