# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  subjectName?: string;
  subjectCode?: string;
  department?: string;
  semester?: string;
  /** @nullable */
  facultyId?: number | null;
}